requirements.txt
# Core ML/DL
torch
pytorch-lightning

# FastAPI & Web
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2  # Only if using HTML templates

# Data Handling
pandas
numpy
yfinance
matplotlib
seaborn
scikit-learn

# MLOps Tools
mlflow==2.2.2
dvc

# Utilities
python-dotenv
python-box==6.0.2
ensure==1.0.2
tqdm
pyyaml
joblib

# Dev/Testing
pytest
httpx

-e .